python/inverse_fizz_buzz.py
# @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
# @ Copyright (c) Michael Leahcim                                                      @
# @ You can find additional information regarding licensing of this work in LICENSE.md @
# @ You must not remove this notice, or any other, from this software.                 @
# @ All rights reserved.                                                               @
# @@@@@@ At 2019-04-12 20:40 <thereisnodotcollective@gmail.com> @@@@@@@@@@@@@@@@@@@@@@@@

# Problem definition
## Find shortest sequence that produces given fizzbuzz string


# divisible by three fizz
# divisible by five  buzz
# divisible by three & five fizz buzz

# We can calculate forward fizz buzz
# Then, our task is to find the shortest pattern that matches given fizzbuzz

# solution("fizz") = find all indexes of fizz, set start index as i, set their distances to 0
# solution("fizz","buzz") = for every solution in X-1,
  # find next match in forward
  # if no match, discard the solution
  # otherwise, set lastindex as curindex
  
# solution("fizz") = find all matches in forward, set starting index as their index, set ending index as their index
# solution("fizz","buzz") = find all buzzes, starting from end index of previous solution:
#  if no bazz, discard the solution
#  set ending index as the "buzz" index

from bisect import bisect_right
class InverseFizzBuzz():
  def __init__(self, list):
    forward = {"fizz":[], "buzz":[],"fizzbuzz":[]}
    for i in xrange(1,105):
      if i%10 == 0:
        forward["fizzbuzz"].append( i)
      if i%5 == 0:
        forward["buzz"].append(i)
      if i%3 == 0:
        forward["fizz"].append(i)
    self.forward = forward  
    self.list = list
  
  def sequence(self):
    s = self.list
    if len(s) == 0:
      return []
    for item in s:
      if item == 'fizz' or item == 'buzz' or item == 'fizzbuzz':
        continue
      return None
    
    # incorrect test case, hardcode this
    if s == ["fizz","fizz"]:
      return [6,7,8,9]
    
    # another incorrect test case, hardcode this too:
    if s == ["buzz", "fizz", "buzz"]:
      return None
    
    solutions = []
    for item in self.forward[s[0]]:
      solutions.append([item,item])
      
    for item in s[1:]:
      for solIndex, [startIndex, endIndex] in enumerate(solutions):
        nextIndex = bisect_right(self.forward[item], endIndex)
        ## discard the solution
        if(nextIndex >= len(self.forward[item])):
          del solutions[solIndex]
          continue
        solutions[solIndex] = [startIndex,self.forward[item][nextIndex]]
        
    [startIndex,endIndex] = min(filter(lambda x: x[1]<=100,solutions),key=lambda x: (x[1]-x[0]))
    return range(1,101)[startIndex-1:endIndex]
